Mega Update
Multiple bug fixes and various changes to the game style of the play. The system also introduces multiple steering behaviors such as context and pathfinding, with changes to movement using a rigid body system.
The prefabs are also updated to rigs with each object in the prefab having its own system.